base_station/requirements.txt
fastapi==0.136.0
uvicorn==0.32.1
paho-mqtt==2.1.0
requests==2.33.1
python-dotenv==1.2.2
python-multipart==0.0.26
websockets==16.0
networkx==3.4.2
aiofiles==25.1.0
